// ts/enums.d.ts
export const enum CardType {
	MINION = 4,
	INVALID = 0,
	HERO = 3,
	TOKEN = 9,
	SPELL = 5,
	ENCHANTMENT = 6,
	HERO_POWER = 10,
	ABILITY = 5,
	PLAYER = 2,
	ITEM = 8,
	GAME = 1,
	WEAPON = 7
}

export const enum PlayState {
	INVALID = 0,
	CONCEDED = 8,
	DISCONNECTED = 7,
	LOSING = 3,
	WINNING = 2,
	TIED = 6,
	LOST = 5,
	WON = 4,
	QUIT = 8,
	PLAYING = 1
}

export const enum ChoiceType {
	INVALID = 0,
	MULLIGAN = 1,
	GENERAL = 2
}

export const enum BnetRegion {
	REGION_UNKNOWN = 0,
	REGION_US = 1,
	REGION_LIVE_VERIFICATION = 40,
	REGION_CN = 5,
	REGION_MSCHWEITZER_BN11 = 52,
	REGION_TW = 4,
	REGION_UNINITIALIZED = -1,
	REGION_KR = 3,
	REGION_PTR_LOC = 41,
	REGION_MSCHWEITZER_BN12 = 53,
	REGION_EU = 2,
	REGION_DEV = 60,
	REGION_PTR = 98
}

export const enum MetaDataType {
	META_HEALING = 2,
	META_DAMAGE = 1,
	DAMAGE = 1,
	META_TARGET = 0,
	JOUST = 3,
	TARGET = 0,
	CLIENT_HISTORY = 4,
	HEALING = 2
}

export const enum Type {
	STRING = 8,
	COUNTER = 3,
	ENTITY = 4,
	LOCSTRING = -2,
	UNKNOWN = 0,
	NUMBER = 2,
	TEAM = 6,
	BOOL = 1,
	PLAYER = 5,
	ENTITY_DEFINITION = 7
}

export const enum DeckType {
	TAVERN_BRAWL_DECK = 6,
	AI_DECK = 2,
	DRAFT_DECK = 4,
	NORMAL_DECK = 1,
	HIDDEN_DECK = 1000,
	PRECON_DECK = 5
}

export const enum Mulligan {
	INPUT = 1,
	INVALID = 0,
	DEALING = 2,
	WAITING = 3,
	DONE = 4
}
// ...
export const enum BlockType {
	DEATHS = 6,
	POWER = 3,
	ACTION = 99,
	SCRIPT = 4,
	ATTACK = 1,
	JOUST = 2,
	TRIGGER = 5,
	CONTINUOUS = 2,
	PLAY = 7,
	FATIGUE = 8,
	RITUAL = 9
}

export const enum FormatType {
	FT_UNKNOWN = 0,
	FT_WILD = 1,
	FT_STANDARD = 2
}

export const enum Rarity {
	LEGENDARY = 5,
	EPIC = 4,
	INVALID = 0,
	UNKNOWN_6 = 6,
	RARE = 3,
	COMMON = 1,
	FREE = 2
}

export const enum Zone {
	DISCARD = -2,
	INVALID = 0,
	SETASIDE = 6,
	HAND = 3,
	DECK = 2,
	REMOVEDFROMGAME = 5,
	PLAY = 1,
	GRAVEYARD = 4,
	SECRET = 7
}

export const enum Race {
	TOTEM = 21,
	DEMON = 15,
	TROLL = 10,
	DRAGON = 24,
	BEAST = 20,
	WORGEN = 12,
	ORC = 8,
	NIGHTELF = 7,
	SCOURGE = 16,
	PIRATE = 23,
	MECHANICAL = 17,
	GOBLIN2 = 13,
	MURLOC = 14,
	INVALID = 0,
	HUMAN = 6,
	OGRE = 19,
	GNOME = 4,
	ELEMENTAL = 18,
	GOBLIN = 5,
	TAUREN = 9,
	BLOODELF = 1,
	DRAENEI = 2,
	NERUBIAN = 22,
	PET = 20,
	UNDEAD = 11,
	DWARF = 3
}

export const enum Step {
	FINAL_WRAPUP = 14,
	INVALID = 0,
	MAIN_BEGIN = 5,
	MAIN_CLEANUP = 16,
	BEGIN_DRAW = 3,
	MAIN_NEXT = 13,
	MAIN_READY = 6,
	MAIN_ACTION = 10,
	BEGIN_FIRST = 1,
	MAIN_DRAW = 8,
	FINAL_GAMEOVER = 15,
	MAIN_START_TRIGGERS = 17,
	MAIN_RESOURCE = 7,
	MAIN_START = 9,
	BEGIN_MULLIGAN = 4,
	BEGIN_SHUFFLE = 2,
	MAIN_COMBAT = 11,
	MAIN_END = 12
}